Split tokens plugin in files and use base-pluign
