+++ public/main/plugins/tokens/erc20-events.js
@@ -1,0 +1,13 @@
+/* eslint-disable max-len */
+
+'use strict'
+
+const erc20Events = [{
+  name: 'Transfer',
+  signature: '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef'
+}, {
+  name: 'Approval',
+  signature: '0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925'
+}]
+
+module.exports = erc20Events
