+++ public/main/plugins/tokens/index.js
@@ -1,219 +1,82 @@
 'use strict'
 
-const abi = require('human-standard-token-abi')
-const logger = require('electron-log')
+const createBasePlugin = require('../../base-plugin')
 
-const WalletError = require('../../WalletError')
+const createApi = require('./api')
+const sendBalances = require('./balances')
+const transactionParser = require('./transactionParser')
 
-let ethWallet
+const start = (eventsBus, ethWallet) => function (pluginEmitter) {
+  eventsBus.on(
+    'wallet-opened',
+    function ({ walletId, addresses, webContents }) {
+      sendBalances({ ethWallet, walletId, addresses, webContents })
 
-const {
-  getTokenBalance,
-  getTokenContractAddresses,
-  getTokenSymbol,
-  setTokenBalance
-} = require('./settings')
-
-const {
-  erc20Events,
-  topicToAddress,
-  transactionParser
-} = require('./transactionParser')
-
-function sendBalances ({ walletId, addresses, webContents }) {
-  const contractAddresses = getTokenContractAddresses()
-
-  const web3 = ethWallet.getWeb3()
-  const contracts = contractAddresses
-    .map(a => a.toLowerCase())
-    .map(address => ({
-      contractAddress: address,
-      contract: new web3.eth.Contract(abi, address),
-      symbol: getTokenSymbol(address)
-    }))
-
-  addresses.map(a => a.toLowerCase()).forEach(function (address) {
-    contracts.forEach(function ({ contractAddress, contract, symbol }) {
-      contract.methods
-        .balanceOf(address)
-        .call()
-        .then(function (balance) {
-          setTokenBalance({ walletId, address, contractAddress, balance })
-
-          webContents.send('wallet-state-changed', {
-            [walletId]: {
-              addresses: {
-                [address]: {
-                  token: {
-                    [contractAddress]: {
-                      symbol,
-                      balance
-                    }
-                  }
-                }
-              }
-            }
-          })
-          logger.verbose(`<-- ${symbol} ${address} ${balance}`)
-        })
-        .catch(function (err) {
-          logger.warn('Could not get token balance', symbol, err)
-
-          // TODO retry before notifying
-          webContents.send('connectivity-state-changed', {
-            ok: false,
-            reason: 'Connection to Ethereum node failed',
-            plugin: 'tokens',
-            err: err.message
-          })
-
-          // Send cached balance
-          webContents.send('wallet-state-changed', {
-            [walletId]: {
-              addresses: {
-                [address]: {
-                  token: {
-                    [contractAddress]: {
-                      symbol,
-                      balance: getTokenBalance({
-                        walletId,
-                        address,
-                        contractAddress
-                      })
-                    }
-                  }
-                }
-              }
-            }
-          })
-        })
-    })
-  })
-}
-
-let subscriptions = []
-
-function unsubscribeUpdates (_, webContents) {
-  subscriptions = subscriptions.filter(s => s.webContents !== webContents)
-}
-
-function attachToEvents (bus) {
-  bus.on('wallet-opened', function ({ walletId, addresses, webContents }) {
-    sendBalances({ walletId, addresses, webContents })
-
-    webContents.on('destroyed', function () {
-      unsubscribeUpdates(null, webContents)
-    })
-
-    subscriptions = subscriptions.concat({ walletId, addresses, webContents })
-  })
-
-  bus.on('tok-tx-confirmed', function () {
-    subscriptions.forEach(sendBalances)
-  })
-
-  bus.on('tok-tx-unconfirmed', function () {
-    subscriptions.forEach(sendBalances)
-  })
-}
-
-function callTokenMethod (method, args, waitForReceipt) {
-  const { password, token, from, to, value, gasPrice, gasLimit } = args
-
-  logger.verbose(`Calling ${method} of ERC20 token`, {
-    from,
-    to,
-    value,
-    token,
-    gasLimit,
-    gasPrice
-  })
-
-  const web3 = ethWallet.getWeb3()
-  const contract = new web3.eth.Contract(abi, token)
-  const call = contract.methods[method](to, value)
-  const data = call.encodeABI()
-
-  return ethWallet.sendTransaction({ password, from, to: token, data, gasPrice, gasLimit }, waitForReceipt)
-    .then(function (result) {
-      if (!waitForReceipt) {
-        return result
-      }
-
-      const eventName = {
-        transfer: 'Transfer',
-        approve: 'Approval'
-      }[method]
-
-      const signature = erc20Events.find(e => e.name === eventName).signature
-      const success = (result.status === 0 ||
-        result.logs.find(log =>
-          log.address.toLowerCase() === token &&
-          log.topics[0] === signature &&
-          topicToAddress(log.topics[1]) === from.toLowerCase() &&
-          topicToAddress(log.topics[2]) === to.toLowerCase()
-          // TODO validate data === value
-        )
+      pluginEmitter.emit(
+        'register-webcontents-metadata',
+        { webContents, meta: { walletId, addresses } }
       )
+    }
+  )
 
-      if (!success) {
-        throw new WalletError(`Token call ${method} failed`)
-      }
-
-      return result
-    })
-}
-
-function sendToken (args, waitForReceipt) {
-  return callTokenMethod('transfer', args, waitForReceipt)
-}
-
-function approveToken (args, waitForReceipt) {
-  return callTokenMethod('approve', args, waitForReceipt)
-}
-
-function getAllowance ({ token, from, to }) {
-  const web3 = ethWallet.getWeb3()
-  const contract = new web3.eth.Contract(abi, token)
-  return contract.methods.allowance(from, to).call()
-}
-
-function getGasLimit ({ token, to, from, value }) {
-  const symbol = getTokenSymbol(token)
-
-  logger.verbose('Getting token gas limit', { to, value, symbol })
-
-  const web3 = ethWallet.getWeb3()
-  const contract = new web3.eth.Contract(abi, token)
-  const transfer = contract.methods.transfer(to, value)
-
-  return transfer.estimateGas({ from }).then(function (gasLimit) {
-    logger.verbose('Token gas limit retrieved', gasLimit)
-
-    return { gasLimit }
+  eventsBus.on('tok-tx-confirmed', function () {
+    pluginEmitter.emit('token-event')
   })
+
+  eventsBus.on('tok-tx-unconfirmed', function () {
+    pluginEmitter.emit('token-event')
+  })
+}
+
+const broadcastBalances = ethWallet => function (subscriptions) {
+  subscriptions.forEach(function ({ webContents, meta }) {
+    const { walletId, addresses } = meta
+
+    sendBalances({ ethWallet, walletId, addresses, webContents })
+  })
+}
+
+const stop = eventsBus => function () {
+  eventsBus.off('wallet-opened')
+
+  eventsBus.off('tok-tx-confirmed')
+
+  eventsBus.off('tok-tx-unconfirmed')
 }
 
 function init ({ plugins, eventsBus }) {
-  ethWallet = plugins.ethWallet
+  const { ethWallet } = plugins
+
+  const {
+    approveToken,
+    getAllowance,
+    getGasLimit,
+    sendToken
+  } = createApi(ethWallet)
 
   ethWallet.registerTxParser(transactionParser(ethWallet))
 
-  attachToEvents(eventsBus)
+  const plugin = createBasePlugin({
+    start: start(eventsBus, ethWallet),
+    stop: stop(eventsBus),
+    onPluginEvents: [{
+      eventName: 'token-event',
+      handler: broadcastBalances(ethWallet)
+    }]
+  })
 
-  return {
-    name: 'tokens',
-    api: {
-      approveToken,
-      getAllowance
-    },
-    dependencies: ['ethWallet'],
-    uiHooks: [
-      { eventName: 'send-token', auth: true, handler: args => sendToken(args) },
-      { eventName: 'ui-unload', handler: unsubscribeUpdates },
-      { eventName: 'tokens-get-gas-limit', handler: getGasLimit }
-    ]
+  plugin.name = 'tokens'
+  plugin.api = {
+    approveToken,
+    getAllowance
   }
+  plugin.dependencies = ['ethWallet']
+  plugin.uiHooks.push(...[
+    { eventName: 'send-token', auth: true, handler: args => sendToken(args) },
+    { eventName: 'tokens-get-gas-limit', handler: getGasLimit }
+  ])
+
+  return plugin
 }
 
 module.exports = { init }
