+++ public/main/plugins/tokens/topic-to-address.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const topicToAddress = topic => `0x${topic.substr(-40)}`.toLowerCase()
+
+module.exports = topicToAddress
